Reorganize process-flow-lib
