+++ package.json
@@ -13,6 +13,7 @@
   },
   "scripts": {
     "start": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng serve --live-reload",
+    "build:process-flow-lib": "tsc -p src/process-flow-lib/tsconfig.json",
     "build": "webpack build --config ./process-flow-diagram-component/webpack-build-mfe.config.js --mode=development && node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --base-href .",
     "build-watch": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --base-href . --watch",
     "build-diagram": "webpack build --config ./process-flow-diagram-component/webpack-build-mfe.config.js --mode=development --watch",
