+++ src/process-flow-lib/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "compilerOptions": {
+    "outDir": "../../process-flow-diagram-component/dist",
+    "rootDir": ".",
+    "declaration": true,
+    "target": "es2020",
+    "module": "esnext",
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "sourceMap": true,
+    "strict": false,
+  },
+  "include": [
+    "./**/*.ts",
+  ]
+}
