--- src/process-flow-types/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "compilerOptions": {
-      "outDir": "../../process-flow-diagram-component/lib",
-      "rootDir": ".",
-      "declaration": true,
-      "composite": true,
-      "target": "es2020",
-      "module": "esnext",
-      "moduleResolution": "node",
-    },
-  }
